Update Spanish localization
Translated mana translations and updated the descriptions to make them fit in their boxes.
--
I believe there are some missing translations, but don't know for sure how I should format them from the Eng file. By the way, the Normal and Master's titles look too tiny in their boxes (almost unreadable). I'm not sure if this is intended, but their length make them look like that in Spanish. (Maybe it's a known issue but I wanted to comment it).
Thanks!